fix(ai-director): use u32 for max_parallel_files serialization
Change max_parallel_files type from Option<usize> to Option<u32>
for proper serde serialization across Tauri boundary.
Add explicit usize conversion where needed.

+++ src-tauri/src/analysis/services/ai_director.rs
@@ -315,7 +315,7 @@ pub struct AIDirectorConfig {
   pub enable_parallel_processing: bool,
 
   /// 🆕 Phase 3: Максимальное количество параллельных файлов (None = auto, на основе CPU cores)
-  pub max_parallel_files: Option<usize>,
+  pub max_parallel_files: Option<u32>,
 }
 
 impl Default for AIDirectorConfig {
